Move discount calculation into ShoppingCart

+++ src/main/scala/com/github/rojanu/Offer.scala
@@ -4,6 +4,6 @@ trait Offer {
   def toPay(quantity: Int, price: BigDecimal): BigDecimal
 }
 
-case class QuantityOffer(get: Int, payFor: Int) extends Offer {
+case class QuantityOffer(get: Int, payFor: BigDecimal) extends Offer {
   override def toPay(quantity: Int, price: BigDecimal): BigDecimal = (payFor * (quantity / get) + (quantity % get)) * price
 }
